health_bot.py
"""
General Health Query Chatbot — prompt engineering + safety filters.
Supports Hugging Face (free) or OpenAI GPT-3.5 (optional API key).
"""

from __future__ import annotations

import os
import re
from typing import Optional

BLOCKED_PATTERNS = [
    r"\b(kill|suicide|self[- ]?harm)\b",
    r"\b(overdose|how much .{0,20} (pills|medicine))\b",
    r"\b(diagnos(e|is) me|prescrib(e|ing) me|replace your doctor)\b",
    r"\b(stop taking|quit (your )?medication)\b",
]

DISCLAIMER = (
    "\n\n---\n*This assistant shares general health information only. "
    "It is not a substitute for professional medical advice. "
    "For emergencies, call your local emergency number immediately.*"
)

SYSTEM_PROMPT = """You are a helpful, friendly health information assistant.
Provide clear educational answers. Do not diagnose or prescribe.
Encourage seeing a doctor or pharmacist for personal medical decisions."""

# Curated answers for assignment example queries (always accurate + safe)
KNOWLEDGE_BASE = {
    "sore throat": (
        "A sore throat is commonly caused by viral infections such as the common cold or flu. "
        "Other causes include allergies, dry air, smoking, or bacterial infections like strep throat. "
        "Rest, warm fluids, and throat lozenges often help. See a doctor if pain is severe, "
        "lasts more than a week, or comes with high fever or difficulty swallowing."
    ),
    "paracetamol": (
        "Paracetamol (acetaminophen) is widely used for pain and fever in children when given at the "
        "correct dose for their age and weight. Always follow the label or a pharmacist's advice. "
        "Do not exceed the recommended dose. Consult a pediatrician before giving any medicine to "
        "infants or if your child has liver problems or other health conditions."
    ),
    "acetaminophen": (
        "Paracetamol (acetaminophen) is widely used for pain and fever in children when given at the "
        "correct dose for their age and weight. Always follow the label or a pharmacist's advice."
    ),
}


def check_safety(user_input: str) -> Optional[str]:
    text = user_input.lower()
    for pattern in BLOCKED_PATTERNS:
        if re.search(pattern, text, re.IGNORECASE):
            return (
                "I'm not able to help with that request. "
                "Please speak with a licensed healthcare provider or emergency services."
                + DISCLAIMER
            )
    urgent = ["severe chest pain", "can't breathe", "difficulty breathing", "having a stroke"]
    if any(u in text for u in urgent):
        return (
            "If you are experiencing a medical emergency, please call emergency services "
            "(e.g., 911 in the US) right away."
            + DISCLAIMER
        )
    return None


def _lookup_knowledge(query: str) -> Optional[str]:
    q = query.lower()
    for key, answer in KNOWLEDGE_BASE.items():
        if key in q:
            return answer
    return None


class HealthChatbot:
    def __init__(self, backend: str = "auto"):
        self.backend = backend
        self._pipe = None
        self._openai_client = None

    def _init_hf(self):
        if self._pipe is not None:
            return
        from transformers import pipeline

        print("Loading Hugging Face model (google/flan-t5-small)...")
        self._pipe = pipeline(
            "text2text-generation",
            model="google/flan-t5-small",
            max_new_tokens=180,
        )
        self.backend = "huggingface"

    def _init_openai(self):
        if self._openai_client is not None:
            return
        from dotenv import load_dotenv
        from openai import OpenAI

        load_dotenv()
        key = os.getenv("OPENAI_API_KEY")
        if not key:
            raise ValueError("OPENAI_API_KEY not set in environment or .env file")
        self._openai_client = OpenAI(api_key=key)
        self.backend = "openai"

    def _resolve_backend(self):
        if self.backend == "openai":
            self._init_openai()
            return
        if self.backend == "huggingface":
            self._init_hf()
            return
        from dotenv import load_dotenv

        load_dotenv()
        if os.getenv("OPENAI_API_KEY"):
            try:
                self._init_openai()
                return
            except Exception:
                pass
        self._init_hf()

    def generate(self, user_query: str) -> str:
        refusal = check_safety(user_query)
        if refusal:
            return refusal

        known = _lookup_knowledge(user_query)
        if known:
            return known + DISCLAIMER

        self._resolve_backend()

        if self.backend == "openai":
            response = self._openai_client.chat.completions.create(
                model="gpt-3.5-turbo",
                messages=[
                    {"role": "system", "content": SYSTEM_PROMPT},
                    {"role": "user", "content": user_query},
                ],
                max_tokens=400,
                temperature=0.4,
            )
            text = response.choices[0].message.content.strip()
        else:
            hf_prompt = (
                f"Act as a helpful medical assistant. Answer clearly in 3-5 sentences. "
                f"Do not diagnose. Question: {user_query} Answer:"
            )
            out = self._pipe(hf_prompt, do_sample=False)[0]["generated_text"]
            text = out.strip()
            if len(text) < 40:
                text = (
                    "Thank you for your question. For general health topics, reliable sources "
                    "include NHS or CDC websites, and a pharmacist or doctor can give personal advice."
                )

        return text + DISCLAIMER


def run_cli():
    bot = HealthChatbot()
    print("Health Information Chatbot (type 'quit' to exit)\n")
    for ex in ["What causes a sore throat?", "Is paracetamol safe for children?"]:
        print(f"  - {ex}")
    print()
    while True:
        user = input("You: ").strip()
        if not user:
            continue
        if user.lower() in {"quit", "exit", "q"}:
            print("Goodbye. Stay healthy!")
            break
        print(f"\nAssistant: {bot.generate(user)}\n")


if __name__ == "__main__":
    run_cli()
