requirements.txt
transformers>=4.40.0
torch>=2.0.0
python-dotenv>=1.0.0
openai>=1.30.0
streamlit>=1.33.0
jupyter>=1.0.0
